+++ src/core/leoCommands.ts
@@ -1282,6 +1282,7 @@ export class Commands {
         // This does not prevent zombie windows if the script puts up a dialog...
         try {
             c.inCommand = false;
+            g.app.hasScriptShownlog = false; // Reset this flag. (LeoJS only)
             g.app.inScript = true;
             (g.inScript as boolean) = g.app.inScript; // g.inScript is a synonym for g.app.inScript.