+++ src/core/leoGlobals.ts
@@ -4027,6 +4027,12 @@ export function es(...args: any[]): void {
     });
 
     if (app && app.gui) {
+        // Make sure the log pane is visible with 'show Log Pane' command (LeoJS only)
+        if (app.inScript && !app.hasScriptShownlog) {
+            app.gui.showLogPane();
+            app.hasScriptShownlog = true;
+        }
+
         app.gui.addLogPaneEntry(s);
     } else {
         logBuffer.push(s);